split functions from unsupervised into more files, remove .gitignore from playground
